pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--ignore=tests/test_live_integrations.py"
pythonpath = ["."]

[project]
name = "amanat"
version = "0.1.0"
description = "Privacy-first data governance agent for humanitarian NGOs, powered by Auth0 Token Vault and IBM Granite"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "auth0-ai>=1.0.2",
    "chainlit>=2.10.0",
    "cryptography>=44.0.0",
    "docling>=2.84.0",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.0",
    "rank-bm25>=0.2.2",
    "rich>=14.3.3",
    "strands-agents>=1.34.1",
    "strands-agents-tools>=0.3.0",
]

[project.optional-dependencies]
apple = [
    "mlx>=0.31.1",
    "mlx-lm>=0.31.1",
    "mlx-vlm>=0.4.2",
]
